# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "منصة خبرات - Khibrat HR SaaS API",
        "description": "التوثيق الرسمي لكافة واجهات برمجة التطبيقات (APIs) لمنصة خبرات لإدارة الموارد البشرية",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "سيرفر التطوير المحلي"
        }
    ],
    "paths": {
...
        "/api/companies/{company}/hr-managers": {
            "get": {
                "tags": [
                    "HR Managers"
                ],
                "summary": "List HR managers for a company",
                "operationId": "e02140525938c18f8c8dd51df87c2aad",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of HR managers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "HR Managers"
                ],
                "summary": "Create a new HR manager for a company",
                "operationId": "e8d4293b28ae39e42ff3c442de73fe78",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "email",
                                    "department_id",
                                    "job_title",
                                    "base_salary",
                                    "hire_date"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Sarah Ahmed"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "sarah@company.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+963999888777"
                                    },
                                    "department_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "123e4567-e89b-12d3-a456-426614174000"
                                    },
                                    "employee_code": {
                                        "type": "string",
                                        "example": "HRM-001"
                                    },
                                    "education": {
                                        "type": "string",
                                        "example": "Bachelor of Business Administration"
                                    },
                                    "job_title": {
                                        "type": "string",
                                        "example": "HR Manager"
                                    },
                                    "base_salary": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1200.5
                                    },
                                    "hire_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-01"
                                    },
                                    "employment_type": {
                                        "type": "string",
                                        "example": "full-time"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "HR manager created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "HR manager created successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "hr_manager": {
                                                    "type": "object"
                                                },
                                                "temporary_password": {
                                                    "type": "string",
                                                    "example": "XyZ123!abc"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/companies/{company}/hr-managers/{hr_manager}/activate": {
            "post": {
                "tags": [
                    "HR Managers"
                ],
                "summary": "Activate an HR manager account",
                "operationId": "d8fa2d34c789eb055de3b71df7974529",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "hr_manager",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HR manager activated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "HR manager activated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/companies/{company}/hr-managers/{hr_manager}/deactivate": {
            "post": {
                "tags": [
                    "HR Managers"
                ],
                "summary": "Deactivate an HR manager account",
                "operationId": "c1c1880bbc9e27fff6a1fcabd9de9723",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "hr_manager",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "HR manager deactivated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "HR manager deactivated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "أدخلي الـ Token الذي حصلتِ عليه من الـ Login هنا لتفعيل الصلاحيات",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}